refactor filteringHelpers modularization
