+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "ecomnext",
-  "version": "0.10.3",
+  "version": "0.10.4",
   "private": true,
   "scripts": {
     "dev": "next dev",
